+++ src/FidelityFX.cpp
@@ -35,6 +35,19 @@ void FidelityFX::Init()
 	ffxLoadFunctions(&ffxModule, dll);
 }
 
+void FidelityFX::WrapSwapChain()
+{
+	auto swapChain = DX12SwapChain::GetSingleton();
+
+	ffx::CreateContextDescFrameGenerationSwapChainWrapDX12 desc{};
+	desc.swapchain = swapChain->swapChain.put();
+	desc.gameQueue = swapChain->commandQueue.get();
+
+	ffx::Context swapChainContext{};
+
+	ffx::CreateContext(swapChainContext, nullptr, desc);
+}
+
 void FidelityFX::CreateFrameGenerationResources()
 {
 	auto swapChain = DX12SwapChain::GetSingleton();
@@ -42,14 +55,81 @@ void FidelityFX::CreateFrameGenerationResources()
 	ffx::CreateContextDescFrameGeneration createFg{};
 	createFg.displaySize = { swapChain->swapChainDesc.Width, swapChain->swapChainDesc.Height };
 	createFg.maxRenderSize = createFg.displaySize;
-	createFg.flags = FFX_FRAMEGENERATION_ENABLE_ASYNC_WORKLOAD_SUPPORT;
+	createFg.flags = 0;
 	createFg.backBufferFormat = FFX_API_SURFACE_FORMAT_R8G8B8A8_UNORM;
 
 	ffx::CreateBackendDX12Desc createBackend{};
-	createBackend.header.type = FFX_API_CREATE_CONTEXT_DESC_TYPE_BACKEND_DX12;
 	createBackend.device = swapChain->d3d12Device.get();
 
 	ffx::CreateContext(frameGenContext, nullptr, createFg, createBackend);
+}
+
+void FidelityFX::ConfigureFrameGeneration()
+{
+	// Update frame generation config
+
+	auto upscaling = Upscaling::GetSingleton();
+
+	ffx::ConfigureDescFrameGeneration configParameters{};
+
+	configParameters.frameGenerationEnabled = true;
+	configParameters.flags = 0;
+	//configParameters.flags |= m_DrawFrameGenerationDebugTearLines ? FFX_FRAMEGENERATION_FLAG_DRAW_DEBUG_TEAR_LINES : 0;
+	//configParameters.flags |= m_DrawFrameGenerationDebugResetIndicators ? FFX_FRAMEGENERATION_FLAG_DRAW_DEBUG_RESET_INDICATORS : 0;
+	//configParameters.flags |= m_DrawFrameGenerationDebugView ? FFX_FRAMEGENERATION_FLAG_DRAW_DEBUG_VIEW : 0;
+	configParameters.HUDLessColor = ffxApiGetResourceDX12(upscaling->colorBufferShared12.get(), FFX_API_RESOURCE_STATE_COMPUTE_READ);
+	configParameters.allowAsyncWorkloads = false;
+	// assume symmetric letterbox
+
+	auto swapChain = DX12SwapChain::GetSingleton();
+
+	configParameters.generationRect.left = (swapChain->swapChainDesc.Width - swapChain->swapChainDesc.Width) / 2;
+	configParameters.generationRect.top = (swapChain->swapChainDesc.Height - swapChain->swapChainDesc.Height) / 2;
+	configParameters.generationRect.width = swapChain->swapChainDesc.Width;
+	configParameters.generationRect.height = swapChain->swapChainDesc.Height;
+
+	configParameters.frameGenerationCallback = nullptr;
+	configParameters.frameGenerationCallbackUserContext = nullptr;
+	
+	configParameters.onlyPresentGenerated = false;
+
+	static uint64_t frameID = 0;
+	configParameters.frameID = frameID;
+
+	configParameters.swapChain = swapChain->swapChain.get();
+
+	ffx::Configure(frameGenContext, configParameters);
+
+	ffx::DispatchDescFrameGenerationPrepare dispatchParameters{};
+
+	dispatchParameters.commandList = swapChain->commandList.get();
+
+	dispatchParameters.motionVectorScale.x = (float)swapChain->swapChainDesc.Width;
+	dispatchParameters.motionVectorScale.y = (float)swapChain->swapChainDesc.Height;
+	dispatchParameters.renderSize.width = swapChain->swapChainDesc.Width;
+	dispatchParameters.renderSize.height = swapChain->swapChainDesc.Height;
+	dispatchParameters.jitterOffset.x = 0;
+	dispatchParameters.jitterOffset.y = 0;
+
+	static float& deltaTime = (*(float*)REL::RelocationID(523660, 410199).address());
+	dispatchParameters.frameTimeDelta = deltaTime * 1000.f;
+
+	static float& cameraNear = (*(float*)(REL::RelocationID(517032, 403540).address() + 0x40));
+	static float& cameraFar = (*(float*)(REL::RelocationID(517032, 403540).address() + 0x44));
+	dispatchParameters.cameraFar = cameraFar;
+	dispatchParameters.cameraNear = cameraNear;
+
+	dispatchParameters.cameraFovAngleVertical = Util::GetVerticalFOVRad();
+	dispatchParameters.viewSpaceToMetersFactor = 0.01428222656f;
+
+	dispatchParameters.frameID = frameID;
+
+	dispatchParameters.depth = ffxApiGetResourceDX12(upscaling->depthBufferShared12.get(), FFX_API_RESOURCE_STATE_COMPUTE_READ);
+	dispatchParameters.motionVectors = ffxApiGetResourceDX12(swapChain->renderTargetsD3D12[RE::RENDER_TARGETS::RENDER_TARGET::kMOTION_VECTOR].d3d12Resource.get(), FFX_API_RESOURCE_STATE_COMPUTE_READ);
+	
+	ffx::Dispatch(frameGenContext, dispatchParameters);
+
+	frameID++;
 }
 
 void FidelityFX::CreateFSRResources()
