+++ src/Streamline.h
@@ -105,14 +105,6 @@ public:
 
 	PFun_slPCLSetMarker* slPCLSetMarker2{};
 
-	Texture2D* colorBufferShared;
-	Texture2D* depthBufferShared;
-
-	winrt::com_ptr<ID3D12Resource> colorBufferShared12;
-	winrt::com_ptr<ID3D12Resource> depthBufferShared12;
-
-	ID3D11ComputeShader* copyDepthToSharedBufferCS;
-
 	void DrawSettings();
 
 	void LoadInterposer();
@@ -158,19 +150,8 @@ public:
 		static inline REL::Relocation<decltype(thunk)> func;
 	};
 
-	struct MenuManagerDrawInterfaceStartHook
-	{
-		static void thunk(int64_t a1)
-		{
-			GetSingleton()->CopyResourcesToSharedBuffers();
-			func(a1);
-		}
-		static inline REL::Relocation<decltype(thunk)> func;
-	};
-
 	static void InstallHooks()
 	{
 		stl::write_thunk_call<Main_RenderWorld>(REL::RelocationID(35560, 36559).address() + REL::Relocate(0x831, 0x841, 0x791));
-		stl::write_thunk_call<MenuManagerDrawInterfaceStartHook>(REL::RelocationID(79947, 82084).address() + REL::Relocate(0x7E, 0x83, 0x97));
 	}
 };
