+++ src/Upscaling.h
@@ -77,6 +77,17 @@ public:
 	void CreateUpscalingResources();
 	void DestroyUpscalingResources();
 
+	Texture2D* colorBufferShared;
+	Texture2D* depthBufferShared;
+
+	winrt::com_ptr<ID3D12Resource> colorBufferShared12;
+	winrt::com_ptr<ID3D12Resource> depthBufferShared12;
+
+	ID3D11ComputeShader* copyDepthToSharedBufferCS;
+
+	void CreateFrameGenerationResources();
+	void CopyResourcesToSharedBuffers();
+
 	struct Main_UpdateJitter
 	{
 		static void thunk(RE::BSGraphics::State* a_state)
@@ -128,6 +139,16 @@ public:
 		static inline REL::Relocation<decltype(thunk)> func;
 	};
 
+	struct MenuManagerDrawInterfaceStartHook
+	{
+		static void thunk(int64_t a1)
+		{
+			GetSingleton()->CopyResourcesToSharedBuffers();
+			func(a1);
+		}
+		static inline REL::Relocation<decltype(thunk)> func;
+	};
+
 	static void InstallHooks()
 	{
 		if (!State::GetSingleton()->upscalerLoaded) {
@@ -137,6 +158,8 @@ public:
 			stl::write_thunk_call<TAA_BeginTechnique>(REL::RelocationID(100540, 107270).address() + REL::Relocate(0x3E9, 0x3EA, 0x448));
 			stl::write_thunk_call<TAA_EndTechnique>(REL::RelocationID(100540, 107270).address() + REL::Relocate(0x3F3, 0x3F4, 0x452));
 			stl::write_thunk_call<BSImageSpacerShader_RenderPassImmediately>(REL::RelocationID(100951, 107733).address() + REL::Relocate(0x82, 0x78, 0x7E));
+			
+			stl::write_thunk_call<MenuManagerDrawInterfaceStartHook>(REL::RelocationID(79947, 82084).address() + REL::Relocate(0x7E, 0x83, 0x97));
 
 			logger::info("[Upscaling] Installed hooks");
